Rename files removing unnecessary graphapi_ prefix
