+++ src/graphapi/graphapi.cpp
@@ -24,9 +24,9 @@
  *
  *******************************************************************************/
 #include <miopen/errors.hpp>
+#include <miopen/graphapi/convolution.hpp>
 #include <miopen/graphapi/graphapi.hpp>
-#include <miopen/graphapi/graphapi_convolution.hpp>
-#include <miopen/graphapi/graphapi_tensor.hpp>
+#include <miopen/graphapi/tensor.hpp>
 #include <miopen/logger.hpp>
 
 #include <memory>
