+++ src/include/miopen/graphapi/variant_pack.hpp
